+++ src/Kephas.Validation/Validation/ValidationException.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="DataValidationException.cs" company="Kephas Software SRL">
+// <copyright file="ValidationException.cs" company="Kephas Software SRL">
 //   Copyright (c) Kephas Software SRL. All rights reserved.
 //   Licensed under the MIT license. See LICENSE file in the project root for full license information.
 // </copyright>
@@ -8,24 +8,22 @@
 // </summary>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace Kephas.Data.Validation;
+namespace Kephas.Validation;
 
-using System;
-using System.Linq;
-
+using Kephas.Data;
 using Kephas.Resources;
 
 /// <summary>
 /// Base class for validation exceptions.
 /// </summary>
-public class DataValidationException : DataException
+public class ValidationException : DataException
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="DataValidationException"/> class.
+    /// Initializes a new instance of the <see cref="ValidationException"/> class.
     /// </summary>
     /// <param name="validatedEntity">The validated entity.</param>
     /// <param name="validationResult">The validation result.</param>
-    public DataValidationException(object validatedEntity, IDataValidationResult validationResult)
+    public ValidationException(object validatedEntity, IValidationResult validationResult)
         : base(string.Format(ValidationStrings.DataValidationException_Message, validatedEntity, string.Join(Environment.NewLine, validationResult.Select(e => e.Message))))
     {
         this.ValidatedEntity = validatedEntity;
@@ -33,18 +31,18 @@ public class DataValidationException : DataException
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="DataValidationException"/> class.
+    /// Initializes a new instance of the <see cref="ValidationException"/> class.
     /// </summary>
     /// <param name="message">
     /// The message.
     /// </param>
-    protected DataValidationException(string message)
+    protected ValidationException(string message)
         : base(message)
     {
     }
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="DataValidationException"/> class.
+    /// Initializes a new instance of the <see cref="ValidationException"/> class.
     /// </summary>
     /// <param name="message">
     /// The message.
@@ -52,7 +50,7 @@ public class DataValidationException : DataException
     /// <param name="inner">
     /// The inner.
     /// </param>
-    protected DataValidationException(string message, Exception inner)
+    protected ValidationException(string message, Exception inner)
         : base(message, inner)
     {
     }
@@ -71,5 +69,5 @@ public class DataValidationException : DataException
     /// <value>
     /// The validation result.
     /// </value>
-    public IDataValidationResult? ValidationResult { get; }
+    public IValidationResult? ValidationResult { get; }
 }