+++ src/Kephas.Validation/release-notes.md
@@ -6,5 +6,6 @@
 # 12.0.0
 
 * NEW: Added ```README.md``` and ```release-notes.md```.
-* Breaking change: ```DefaultAppManager``` works now with lazy feature managers and behaviors.
-  * Behavior change: the behaviors at both application and feature level can now store state between calls.
+* Breaking change: renamed classes and interfaces from ```DataValidation*``` to ```Validation*```.
+* Breaking change: moved the classes and interfaces from ```kephas.Data.Validation``` namespace to ```Kephas.Validation``` namespace.
+* Breaking change: ```DataValidationSeverity``` replaced by ```SeverityLevel```.